fix utils and update readme

+++ C12AnCpp/Tools/inc/utils.h
@@ -8,7 +8,10 @@
 namespace tools {
   namespace utils {
 
-    std::tuple<double,double,double> getDISvariables( root::particle *, float, float );
+    // compute Q2, xB, W2
+    // takes as input a scattered electron, the energy of the electron beam 
+    // and the id of the target particle
+    std::tuple<double,double,double> getDISvariables( root::particle *, float, int );
   }
 }
